Resolve 2 conflicts:
- IEfiPixService.cs: keep PUT verb (master)
- limpa-codigo-log.md: keep all entries from both sides

+++ EasyStock.Domain/Events/EntradaEstoqueRegistrada.cs
@@ -2,5 +2,5 @@ using System;
 
 namespace EasyStock.Domain.Events
 {
-    public sealed record EntradaEstoqueRegistrada(Guid EventoId, DateTime OcorridoEm, Guid ItemEstoqueId, Guid ProdutoId, Guid EmpresaId, int Quantidade, string? CodigoLote) : DomainEvent(EventoId, OcorridoEm);
+    public sealed record EntradaEstoqueRegistrada(Guid EventoId, DateTime OcorridoEm, Guid ItemEstoqueId, Guid ProdutoId, Guid EmpresaId, decimal Quantidade, string? CodigoLote) : DomainEvent(EventoId, OcorridoEm);
 }
